Fix multiple UI issues and implement unified card editing
- Bubble Words: Sort existing words A-Z in dialog
- Profile: Fix accuracy calculation to cap at 100%
- Profile: Remove duplicate edit icon from level progress circle
- Dark Mode: Fix achievement and reward title text colors
- Cards Page: Change header to 'Taal Trek' for consistency
- Deck Learning: Fix percentage calculation with proper type conversion
- Unified Card Editing: Implement full-screen editing with deck management
- Haptic Feedback: Add comprehensive haptic service
- Import/Export: Improve error handling and add Gemini guide
- Shuffle Cards: Add customization options and restore Write Your Card
- Memory Game: Randomize card colors to prevent color matching
- True/False: Fix question bias and improve button styling
- Dark Mode: Fix various text visibility issues
- Settings: Remove unused sections and integrate haptic controls

+++ flashcard_app/lib/models/deck.dart
@@ -80,7 +80,7 @@ class Deck {
     
     for (final card in cards) {
       print('🔍 Deck calculation: Card "${card.word}" has ${card.learningPercentage}%');
-      totalPercentage += card.learningPercentage;
+      totalPercentage += card.learningPercentage.toDouble();
     }
     
     final result = totalPercentage / cards.length;
@@ -101,7 +101,7 @@ class Deck {
     
     for (final card in cards) {
       print('🔍 Static Deck calculation: Card "${card.word}" has ${card.learningPercentage}%');
-      totalPercentage += card.learningPercentage;
+      totalPercentage += card.learningPercentage.toDouble();
     }
     
     final result = totalPercentage / cards.length;
